# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "swagger test example",
        "title": "Linhome",
        "contact": {
            "name": "Skyler"
        },
        "version": "1.0"
    },
    "host": "192.168.0.158:9000",
    "basePath": "/v1",
    "paths": {
        "/user/Login": {
            "post": {
                "description": "使用者登入功能",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "登入",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登入成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求失敗",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/Login1": {
            "post": {
                "description": "使用者登入功能",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "登入",
                "parameters": [
                    {
                        "type": "string",
                        "description": "使用者名稱",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密碼",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登入成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求失敗",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}